+++ src/Model/CallbackList.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace amcintosh\FreshBooks\Model;
+
+use Spatie\DataTransferObject\Attributes\CastWith;
+use Spatie\DataTransferObject\Casters\ArrayCaster;
+use amcintosh\FreshBooks\Model\AccountingList;
+use amcintosh\FreshBooks\Model\Callback;
+
+/**
+ * Results of callbacks list call containing list of callbacks and pagination data.
+ *
+ * @package amcintosh\FreshBooks\Model
+ * @link https://www.freshbooks.com/api/webhooks
+ */
+class CallbackList extends AccountingList
+{
+    public const RESPONSE_FIELD = 'callbacks';
+
+    #[CastWith(ArrayCaster::class, itemType: Callback::class)]
+    public array $callbacks;
+}